robot_controller/requirements.txt
PyYAML>=6.0.1
influxdb3-python>=0.17.0